+++ test_flask_ckeditor.py
@@ -150,7 +150,7 @@ class CKEditorTestCase(unittest.TestCase):
         self.assertIn('imageUploadUrl: "/1",', rv)
         self.assertIn('filebrowserUploadUrl: "/1",', rv)
         self.assertIn('filebrowserBrowseUrl: "/2",', rv)
-        self.assertIn('extraPlugins: "1,2,filebrowser,codesnippet",', rv)
+        self.assertIn('extraPlugins: "1,2,filebrowser",', rv)
 
     def test_ckeditor_field(self):
         response = self.client.get('/field')
@@ -194,7 +194,7 @@ class CKEditorTestCase(unittest.TestCase):
         plugins = {'filebrowser', 'lineutils', 'widgetselection',
                    'codesnippet', 'filetools', 'popup', 'widget'}
         for plugin in plugins:
-            url = '/ckeditor/static/basic/plugins/%s/plugin.js' % plugin
+            url = f'/ckeditor/static/basic/plugins/{plugin}/plugin.js'
             response = self.client.get(url)
             response.close()
             self.assertEqual(response.status_code, 200)
@@ -202,7 +202,7 @@ class CKEditorTestCase(unittest.TestCase):
         # standard package plugins
         plugins = {'codesnippet', 'filetools', 'popup', 'widget'}
         for plugin in plugins:
-            url = '/ckeditor/static/standard/plugins/%s/plugin.js' % plugin
+            url = f'/ckeditor/static/standard/plugins/{plugin}/plugin.js'
             response = self.client.get(url)
             response.close()
             self.assertEqual(response.status_code, 200)
@@ -210,7 +210,7 @@ class CKEditorTestCase(unittest.TestCase):
         # full package plugins
         plugins = {'codesnippet', 'filetools', 'popup', 'widget'}
         for plugin in plugins:
-            url = '/ckeditor/static/full/plugins/%s/plugin.js' % plugin
+            url = f'/ckeditor/static/full/plugins/{plugin}/plugin.js'
             response = self.client.get(url)
             response.close()
             self.assertEqual(response.status_code, 200)
@@ -286,3 +286,7 @@ class CKEditorTestCase(unittest.TestCase):
             json.loads(rv.data),
             {'uploaded': 0, 'error': {'message': 'new error message'}}
         )
+
+
+if __name__ == '__main__':
+    unittest.main()