+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 Rlease date: -
 
 - Drop Python 2 and 3.6 support.
+- Remove the deprecated ``codesnippet`` argument in ``ckeditor.config()``.
 
 
 0.4.6